Add files via upload
This program is a Number Guessing Game. It generates a random number between 1 and 100 and asks the player to guess it. The program provides hints by displaying "Lower number please" if the guessed number is too high, or "Higher number please" if it's too low. When the correct number is guessed, the program congratulates the player and displays the total number of attempts made. The game utilizes a loop and a random number generator to provide an engaging guessing experience.