Add files via upload
This C program writes user input to a file named `Your_File_name.txt` from the beginning. It dynamically allocates memory to store the input in a character array within a structure. The user can enter text, which is continuously written to the file until "END" is typed in a new line. The program ensures proper memory management and file handling by freeing allocated memory and closing the file after writing.